Don't compare unsigned with signed integer
Commit e2be166f introduced a change that compares UINT_MAX with
a long int

+++ src/daemon/command.c
@@ -201,7 +201,7 @@ int readcmd(usbdevice* kb, const char* line){
         }
         case DELAY: {
             long int delay;
-            if(sscanf(word, "%ld", &delay) == 1 && 0 <= delay && delay < UINT_MAX) {
+            if(sscanf(word, "%ld", &delay) == 1 && 0 <= delay && delay < (long int)UINT_MAX) {
                 // Add delay of `newdelay` microseconds to macro playback
                 kb->delay = (unsigned int)delay;
             } else if(strcmp(word, "on") == 0) {
